Update Discord-Update.sh
Change temporary directory, because /tmp is the recommended storage location for temporary files that are no longer needed after a reboot. This also has the advantage that no files are left behind if the script is unintentionally aborted. The operating system empties the /tmp folder automatically. Plus point... as of Debian 13, the /tmp folder is mounted as “tmpfs”. This means that the content of /tmp is no longer written to a hard disk, but stored in memory. The code has also been tidied up